refactor: move IMarketingTransactionSource/MarketingTransaction to Application/Contracts
Moves the two ad-platform adapter-port types out of Domain into
Application/Features/MarketingInvoices/Contracts, per arch-review #3660.
No behavioral change; ImportedMarketingTransaction and
IImportedMarketingTransactionRepository remain in Domain.
