+++ backend/src/Anela.Heblo.Application/Features/MarketingInvoices/Contracts/IMarketingTransactionSource.cs
@@ -1,4 +1,4 @@
-namespace Anela.Heblo.Domain.Features.MarketingInvoices;
+namespace Anela.Heblo.Application.Features.MarketingInvoices.Contracts;
 
 public interface IMarketingTransactionSource
 {
