+++ backend/test/Anela.Heblo.Tests/Features/MarketingInvoices/MarketingInvoiceImportServiceTests.cs
@@ -1,3 +1,4 @@
+using Anela.Heblo.Application.Features.MarketingInvoices.Contracts;
 using Anela.Heblo.Application.Features.MarketingInvoices.Services;
 using Anela.Heblo.Domain.Features.MarketingInvoices;
 using Microsoft.Extensions.Logging;
